Refactoring main activity to MVP.
